Move contract logic to eosio.unregd
