+++ eosio.unregd.abi
@@ -1,7 +1,11 @@
 {
-  "____comment": "This file was generated by eosio-abigen. DO NOT EDIT - 2018-06-19T00:53:38",
+  "____comment": "This file was generated by eosio-abigen. DO NOT EDIT - 2018-06-26T08:40:26",
   "version": "eosio::abi/1.0",
-  "types": [],
+  "types": [{
+      "new_type_name": "ethereum_address",
+      "type": "string"
+    }
+  ],
   "structs": [{
       "name": "address",
       "base": "",
@@ -10,21 +14,21 @@
           "type": "uint64"
         },{
           "name": "ethereum_address",
-          "type": "string"
+          "type": "ethereum_address"
         },{
           "name": "balance",
           "type": "asset"
         }
       ]
     },{
-      "name": "used",
+      "name": "add",
       "base": "",
       "fields": [{
-          "name": "id",
-          "type": "uint64"
-        },{
           "name": "ethereum_address",
-          "type": "bytes"
+          "type": "ethereum_address"
+        },{
+          "name": "balance",
+          "type": "asset"
         }
       ]
     },{
@@ -47,6 +51,10 @@
     }
   ],
   "actions": [{
+      "name": "add",
+      "type": "add",
+      "ricardian_contract": ""
+    },{
       "name": "regaccount",
       "type": "regaccount",
       "ricardian_contract": ""
@@ -62,16 +70,6 @@
         "uint64"
       ],
       "type": "address"
-    },{
-      "name": "used",
-      "index_type": "i64",
-      "key_names": [
-        "id"
-      ],
-      "key_types": [
-        "uint64"
-      ],
-      "type": "used"
     }
   ],
   "ricardian_clauses": [],
